+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 - ALSA(process_output): pass `silent=true` to `PCM.try_recover`, so it doesn't write to stderr
+- All error enums are now `Clone`.
 
 # Version 0.15.3 (2024-03-04)